Bug 706091: Encode/Decode ICCOutputColors from utf <-> 8 bit.
When we are passed sICCOutputColors on the command line (and hence
in put_params) it will be utf-8 encoded. We need to ensure that
it is 8bit raw when it goes into the colormanagement so it can
match.

Accordingly, in get_params, we need to reverse from 8 bit raw
back to utf-8 encoded.