Coverity ID 382008 - NULL pointer derefernce
The pointer dereference is in a memcpy, but I believe that when this
happens (pda->base is NULL) then the input data will have been exhausted
and so 'len' the bytes to copy will always be zero.

While I don't know of any C run-time which will cause an actual problem
with this, it seems sensible to avoid it.