pdfwrite - move current_font assignment after error check
Coverity ID 381906 this is actually a false positive from Coverity, we
can't restore the current_font, because we haven't saved it. Not only
that but the conditions under which we save and restore the current
font are different. We only do that for the fallback case and this is
a last ditch (and should never happen) error check.

But it makes sense not to corrupt current_font by assigning an invalid
font to it, so don't do the assignment until after the error check.