GhostPDF - fix hybrid file interpretation
Bug #706065 "Error: missing white space after number (Only happens with dNewPDF"

The hybrid file processing required that the initial xref have a start
of 0 and 0 entries, because I mistakenly thought that was what the spec
says. In fact the spec says that it is 'reasonable' but does not require
it to be the case.

The initial xref in this file has a size of 1 so it failed to be
processed as a hybrid file. Because we prefer the XRefStm to the old
xref table when one is available that meant we failed to read the xref
entirely and fell back to repairing the file. Which was only partially
successful.

This commit rearranges hybrid file processing; we now process the xref
table and any /Prev xrefs before we read the XRefStm. The XRefStm
then fills in any gaps (ore edited objects) with the final correct
offsets. The file now runs to completion.