GhostPDF - avoid supurious numeric overflow warnings
Because of the fact that we assemble an integer as we go when parsing
a number, and only at the end convert it to real if we had a decimal,
it was possible for the integer value to exceed the maximum integer,
and give a warning, when parsing a real.

Avoid that by checking to see if 'real' is true and if it is do not
test the integer value (and don't bother accumulating any more integer
digits).