GhostPDF - Add Subtype to substitute fonts
Noticed while working on a different problem. A Widget annotation with
no Appearance, but a Tx field has a /DA which includes (/Times 10 Tf).
This causes a warning about a missing Subtype.

The problem is that Times isn't defined at the time the annotation is
executed, so we create a substitute font, and to do that we create a
dummy PDF font dictionary. But the dictionary doesn't contain a Subtype
key which is required, hence the warning.

Simple solution; add a Subtype.

At the same time, cure a compiler warning caused by assigning double
precision floating point to integer variables.