Possible fix for SEGV in customer file.
Customer 470 reports that a particular file dies with a SEGV,
but cannot supply the file to us. Memory squeezing logs suggest
that if the "printer_buffer" allocation in gdev_prn_allocate
fails, during a 'reallocate' we then die in a future call.

I suspect that 'the_memory' is being left set to point to freed
memory if the buffer allocation fails.