GhostPDF - Correct main stream parsing
This was noticed while I was working on the N-up problem, the test file
for that problem was the 1.7 PDF Reference and the device was pdfwrite.
After fixing the N-up problem I had tried running the full file, and it
failed badly.

The reason turns out to be that at some time in the distant past (long
before we merged the pdfi branch to master) I wrapped the main stream
in a SubFileDeocde filter, with a 'triler' EODString, before parsing
objects when dereferencing in indirect reference.

The PDF_Reference has an annotation which includes, in the text, the
word 'trailer'. Because it isn't an ObjStm it isn't compressed, and so
the SubFileDecode filter signals EOF when it hits that word, leading to
errors.

This commit removes the SubFileDecode filter, which led on to a couple
of other problems; the offset calculation in pdfi_read_stream_object()
was assuming it was never reading from the main stream, because we
always had a filter in place. That meant the calculation was incorrect
when we are reading from the main stream.

/tests_private/comparefiles/Bug696541.pdf has strings in its Info
dictionary which are invalid, this causes us to read to the end of file.
Previously the SubFileDecode would have prevented us triggering the
'eof' in the main stream but because we aren't using that any more we
were falling foul of it because we would rewind the stream, then the
PDF stream object would still have eof set, so we'd be unable to read
anything from it.

So have pdfi_seek() reset the 'eof' flag if we alter the file position
and it is currently set.

There may be more fallout from this change, the cluster runs clean but
that does not test very many broken files. I'll be testing a moderate
collection of files, including the OSS-fuzz files, over the next few
days and may make more fixes.