Bug 706086: Fix OSS Fuzz clist buffer overrun.
clist_copy_bits can copy small bitmaps into the clist uncompressed.
When it copies them it calculates the size required as being:

 raster * (h-1) + width_bytes

where width_bytes is raster, without the padding at the end.

Thus it saves a byte or two.

The routine we have been using to copy the bitmap in since time
immemorial did not know about this, so copied all the lines with
padding. Most of the time this didn't matter, but just occasionally,
if we get a bitmap that would JUST fit, we can overrun the buffer.