GhostPDF - initialise a block of memory
OSS-fuzz 53758

The file has an NChannel colour space which is broken, one of the ink
names for the Process colourants is neither a name nor a string.

That caused us to exit, but we hadn't initialised all the process_names
in the structure (we allocate a block of char *) to hold them. When we
then freed the colour space it tried to release the allocated names,
which used uninitialised memory and so crashed.

Memset the whole block of pointers to NULL before we start to avoid
this.