GhostPDF - honour STOPONERROR
Noticed while looking at \tests\pdf\annot-fit-bbox-in-rect.pdf. A Form
for an annotation AP (appearance) stream is missing the required Subtype
key.

When PDFSTOPONERROR is set this should abandon the processing, but it
didn't because we normally ignore the error.

Added code to check if stoponerror is set and return the error code if
it is.

Also, pdfi_process_one_page() was only preserving error codes from
pdfi_do_annotations() and pdfi_do_acroform() if the previous code was
greater than 0, that should be greater than or equal to 0.

Together these make PDFSTOPONERROR work properly with this error.