Bug 706083: Accelerate heavily clipped axial shadings.
Cull a patch by repeated bisection as often as possible before
passing it on to the tensor patch rendering routines.

While the left/right/top/bottom halves of the patch are completely
outside the clipping area, discard them.

It's still possible to concoct a patch that will fall foul of
this (have a huge patch, centred on the clipping area), but this
solves the problem in the bug.