Erasepage optimisation - prevent SEGV with display device on Linux
The erasepage optimisation device uses the fill_rectangle method from
the terminating device to do operations like actually erase the page.

It turns out the display device (on Linux at least) doesn't actually
implement this method (it may do so with a gtk+ build), leaving it at
NULL, and causing a seg fault if we try to use it.

Guard that here to prevent a crash.