GhostPDF - don't record errors/warnings while repairing a PDF
Its quite possible for us to get garbage while repairing a PDF file
without that indicating an actual (additional) problem in the PDF file.
So don't record any warnings or errors encountered while we are
repairing a PDF file, if any of them are real they will be reported as
we process the file.