GhostPDF - prevent a warning when getting bbox of text
To find the bounding box of text (for positioning it with annotation
appearance generation) we set up the bbox device and draw the text,
using pdfi_show() to draw the string.

But we didn't call BT, nor did we set the text BlockDepth, which caused
the show routines to indicate an error. Fix that by incrementing and
decrementing the block count round the show operation.