GhostPDF - permit the Page dictionary to actually be a stream
We differentiate between streams and dictionaries which some consumers
apparently do not do. When looking at the pdfa.org 'safedocs' collection
the file Dialect-DictIsStream.pdf threw an error.

I don't plan to change all the places that we differentiate between
streams and dictionaries (there are more than 200) unless it should
become apparent that Acrobat does so, but I'll do this one. Should we
later choose to do more cases then this makes a useful example of how
we could do so.

We now process the dictionary and log an error on exit.