pdfwrite - fix error in Linearizing output files
Bug #706116 "Segmentation fault when processing PDF with many vectors"

There are two problems here; first we should not crash, crashing is bad.
Fix that by initialising the PageHints and SharedHints to NULL and not
freeing the individual hints in the PageHints array if the array is
NULL.

The crash occurred because we got an ioerror from rewrite_object() and
that shouldn't happen either. This turns out to be because of an
optimisation. We expand the scratch buffer size as needed, and we use
that same buffer when writing out the modified object. But we were
always decreasing the amount left to write by the original size of the
buffer (16KB) even though we were writing > 512KB at a time.

Obviously this only happens with insanely large Form XObjects, 16KB is
normally enough to write any reasonable Form dictionary, and we only
expand the size if we need to write a larger dictionary.