GhostPDF - Apply optional content dictionaries in XObjects.
This started off from the pdfa.org 'safedocs' file utf16LE-test.pdf

The file has two XObjects, both of which have optional content
dictionaries which are initally 'OFF'. We were not applying the OC
dictionaries for form or image XObjects which meant that we were
drawing the XObjects when we should not.

After that, two test files showed differences;
/tests_private/comparefiles/BFAX35P5.pdf
/tests_private/pdf/sumatra/1348_-_support_Additional_Actions.pdf

I had noticed BFAX35P5 during the deveopment of pdfi but because
Acrobat draws the gray watermark I had assumed this was a progression
when in fact it was a regression. We should not draw the watermark
when we are being a printer (Printed=true). The same is true for
the Sumatra file.

Added a couple of new error types, and fixed a spelling error in an
existing one.