GhostPDL - Don't store references on the stack
OSS-fuzz bug #54046

The problem is that we were leaving a dictionary on the stack and just
using it directly from there. That's because the code is old and that
was acceptable; since we added the repair code it is not allowed
because the repair will flush the stack.

So rework the code to take a reference to the object so that if it
disappears off the stack it doesn't matter. Since we're making changes
anyway, recast the code so it uses the 'on error goto' form that we use
elsewhere, this makes it much easier to do cleanup properly.

Finally remove a pointless (but harmless) assignment that got left
lying around by a previous change.