GhostPDF - improve warnings when dealing with XObjects
Test file Bug688485.pdf

If a SMask is missing the Group (/G) raise an error rather than
a dmprintf.

Same for a missing or unrecognised Subtype (/S)

And for XObjects with the /PS (PostScript) subtype.

Errors and warnings are preferred to printf because we only report them
once at the end rather than on each occurrence which can be annoyingly
verbose.