GhostPDF - fix spurious circular reference error in annotations
This seems to have crept in as a fix for OSS-fuzz bug 51034, going by
the comments. Unfirtunately, it is incorrect. By the time we get to the
point of testing the 'Parent' object number, we might (and probably will)
have dereferenced 'Parent', which will add it to the loop detection, and
will therefore trigger the circular reference detection.

Removing the code solves the problem and does not re-introduce a
circularity problem with OSS-fuzz 51034, and we still do detect the
actual genuine circular reference in the file.

Checked with all the OSS-fuzz files to ensure the altered code doesn't
miss a circular reference by some other means and it seems not.

So just remove it. At the same time, remove some code that was #if 0
out. I can't remember what that was for but it seems we don't need it.