GhostPDF - add some warnings to approximate the old interpreter
If the %PDF-x.y version has non-white space characters between the
minor version number and the EOL the old interpreter issued a warning,
we now do similarly.

If an annotation entry in the Page dictionary was not a dictionary the
old interpreter raised a warning, we were silently ignoring it.