GhostPDF - more 'emulate the old interpreter' warnings
d1_with_color_ops.pdf

Add a warning when a colour operator is used inside a type 3 font
CharProc which uses 'd1' (uncoloured). The spec says we should ignore
the operation, which we do, but the old interpreter also raised a
warning.


DML_003-2010-1_8.pdf

Add a warning when the BBox of a Form has zero width or height (and so
doesn't draw anything because we clip to the BBox).