Bug 706195: CIDType 2 - Avoid interpreting entire glyph for widths
in pdfi, for CIDType 2 fonts, we were using the default glyph_info method to
retrieve a default set of glyph width metrics. Since this relies on interpreting
the entire TTF glyph description, it is relatively slow. And since, in a valid
PDF, we actually replace those width metrics with values from the W/W2 arrays,
this was a waste.

We'll now only call the default glyph_info method for those width metrics if we
haven't been able to retrieve them from the W/W2 arrays.

We do still call the glyph_info to retrieve any other information requested, but
if we already have the width values, we remove those flags from the request
before passing it on.