GhostPDF - fix %PDF header parsing
My test file "MedTerm Flash Cards2.pdf"

The file has garbage prepended, and the garbage is binary which contains
NULLs. This is confusing the simple-minded 'strstr' C string parsing
because it fails to find a header.

Rewrite the parsing so that it continues checking if there are bytes
available and we haven't found a %PDF yet.

Add a new warning to indicate garbage in front of the header when this
occurs.