GhostPDF - detect circular references when checking Resources for transparency
My test file "Pages from Saddle_reduced.pdf"

The file has a circular reference in objects 1663 and 1664:

1663 0 obj<</XObject<</I0 1664 0 R>>/ProcSet[/PDF/ImageB/ImageC/ImageI/Text]>>endobj
1664 0 obj<</Length 0/Type/XObject/BBox[0 0 684 864]/Resources 1663 0 R/Subtype/Form/FormType 1/Matrix[1 0 0 1 0 0]>>

When retrieving objects from a dictionary we mark/clear the loop
detection around any required dereference, which means we don't detect
circular references.

This seems wrong to me, but when I tried to change it lots of files
started generating errors.....

The whole circular reference mechanism is somewhat flaky and probably
needs a rethink. In the meantime we can detect the error by manually
adding the Resources dictionary to the loop detection code before we
recursively check the dictionary.