GhostPDF - do not abort on ExtGState errors (unless PDFSTOPONERROR)
My test file PATTYP1-red.pdf but also tests_private/pdf/PDFIA1.7_SUBSET/IA3Z0970.pdf

If an ExtGState contains a BG or UCR key and the value is a name
rather than a function, then instead of exiting the ExtGState with an
error log the error as normal and carry on.

Similarly if a TR has a value which is a name other than /Identity.