GhostPDF - Be prepared for fonts to be fonts, not dictionaries
tests_private/pdf/sumatra/x_-_maybe_crashes_under_XP.pdf

When we turn a PDF font dictionary into a PDF font we replace the cached
object in the cache. This means that code retrieving the font after
that can get back a PDF font object not a PDF dictionary.

The code for checking fonts for transparency wasn't expecting that. In
addition a logic flaw in the loop meant that any non-dictionary objects
we encountered were left on the 'loop detection' stack.

Taken together these were causing spurious circular reference errors,
while at the same time not properly checking fonts for transparency
(if they had previously been dereferenced and cached).

I've reworked the code so that it handled PDF font objects if they are
retrieved from the Resource dictioanry, and properly checks them,
correctly removes objects (including objects of the wrong type) from the
loop detection stack, and flags a warning if a Font Resource is of the
wrong type (neither dictionary nor font).

A Font of the wrong type in the Resources dictionary is not necessarily
an error, the Page may never actually use the named font resource.