GhostPDF - avoid errors when running transparency groups for text
tests_private/pdf/sumatra/x_-_recursive_xobjects_cause_crash.pdf

When setting up transparency for a text operation we may need to
execute a Mask Group. That Group is permitted to execute operations that
would not normally be permitted whie drawing text because, obviously,
it affects the Group, not the text.

We were inappropriately warning about the use of invalid operations
inside a text block.

Fix by (ugly) setting the text BlockDepth to 0 (because being non-zero
is the indication to the various operators that we are in a text block)
for the duration of the mask group only.