GhostPDF - give a warning for some kinds of broken font
My test file zero-length-font.pdf

The file has a type 1 font with a FontFile3 where the stream has a
/Length of 0.

The interpreter properly substitutes a font but doesn't mention the
problem with the embedded font, which the old interpreter did. This just
duplicates the code directly above if pdfi_load_font_buffer() fails and
gives a more helpful warning.