GhostPDF - improve Resource finding.
The code was exiting early if we received an undefined error from
dictionary retrieval, which meant it wasn't falling back as far as it
should.

Also the error about inheriting Resources from an earlier stream was
simply wrong, because it was triggered on an error. The noted test
file 'tests_private/comparefiles/Bug692783.pdf' does not inherit in
that fashion, the Resources are genuinely undefined in the file.

These changes mean we fall back further and give an error if we
completely fail to find a named Resource.