GhostPDF - fix a signed/unsigned mismatch
OSS-fuzz #54412

There are several places in read_xref_stream_entries() where we read
a number of bytes from a file and if we got less than expected we
return an error. However the number expected was being passed as an
unsigned int 64 whereas the number of bytes was a signed 64-bit int.

The result of this is that negative numbers were treated incorrectly,
and that's what we get back from pdfi_read_bytes() at EOF.

So change the parameter passing from uint64_t to int64_t, we aren't
really going to be worrying about reading that many bytes. For some
reason the code was taking much longer to complete on Linux than on
Windows, this makes it exit much faster, as indeed it should.