GhostPDF - better detection of invalid Resource inheritance
Bug693490.pdf

We were not detecting (or warning) that an XObject used resources which
were not defined in the XObjcet, nor the current page, but the XObject
from which the nested XObject is executed.

XObjects don't normally contain a /Parent key, but we add one when we
execute a Form XObject, precisely for this file and others like it. But
we were not warning when we exercised this fallback. Note the fact that
we are checking an XObject and if we later find the needed Resource in
the /Parent dictionary then give a warning.