GhostPDF - only set a 'bad pattern' warning for a lack of Resources
infinite_pattern_recursion.pdf

According to the spec /Resources is a required entry in a Pattern, but
this file has a Pattern which doesn't have one (or need one).

But we should only set the error code to 0 if it is 'undefined' when
we try to retrieve the Resources. Other errors sbhould still be errors.