GhostPDF - PDF 2.0, Halftones can be arrays
/tests_private/pdf/PDF_2.0_FTS/fts_24_2432.pdf

Missed this somewhere along the way; in PDF 2.0 a /HT (halftone) can
have an array value which is an array of spot names, the consumer is
supposed to use the first one it recognises.....

We weren't expecting an array and returned a typecheck error. This was
masked by the fact that some of the halftones in the test file
genuinely don't contain usable halftone names so an error is to be
expected.

Add code to deal with an array by trying to find each spot name in turn
in the list of spot names we know about.

Causes diffs in /tests_private/pdf/PDF_2.0_FTS/fts_24_2432.pdf because
we now use a different spot shape in two cases.