GhostPDF - tighten the search for a header
A test file (which I've no lost track of) had comments before the
PDF header which included the text "%PDF" which we were incorrectly
identifying as being the start of the header and causing a spurious
error about the version number.

There's no way to completely avoid someone putting what appears to be
a genuine PDF header in comments, but tightening the requirement makes
it less likely we'll be tripped up.