GhostPDF - avoid supurious errors on XRefStm
If a startxref was incorrect, but happened to point to an integer, we
would assume that it was pointing to an XRefStm rather than assuming
the startxref was wrong.

We would then try to read an XRefStm, fail, and raise an error about
the XRefStm being incorrect, in a file which does not contain an
XRefStm!

Modify the code so that in this case we check to see if what we have
been pointed to is a real object and if it isn't assume the startxref
is incorrect an rebuild the xref.

When we do have a /XRefStm (in a hybrid file), carry on as before.