GhostPDF - clear stack after colour operations inside type 3 'd1'
"z:\tests\pdf2.0\pdf2ftsRel2\fts_23_2310.pdf"

If a Type 3 font has a CharProc which uses the 'd1' operator to declare
that it does not set colour, but then goes on to set the colour space
or colour, we raise an error (correctly).

But I'd missed the cs, CS, sc, SC, scn and SCN operators.

So firstly add code to ignore the colour space change for cs and CS.

Secondly ignore the actual colour from sc, SC, scn and SCN. This is a
little more difficult. Because we haven't set the colour space, we
don't know how many components should be on the stack. So we just
clear the stack wholesale. The PDF file is already broken, and the
stack ought to be clear anyway.

This was wrongly causing a 'garbage left on stack' error.