GhostPDF - implement PDF 2.0 feature Polyline with Path
My test file pdf2ftsRel2\fts_32_3252.pdf

We don't seem to have the PDF 2.0 functional tests in the cluster suite
so this doesn't seem to get tested. PDF 2.0 added a new entry to
Polyline annotations 'Path' instead of 'Vertices', allowing for curves.

This commit adds support for that; I'm not entirely convinced about the
line ending code but I don't have a PDF 2.0 capable Acrobat to check
against so I'm going to go with this until someone complains.

There are no cluster diffs because the annotation has a /Flags of 0 and
so is not rendered on printers.