GhostPDF - control emission of Boxes from the pdfwrite device
Bug #706128 "Scaling, translating, and writing text does not work properly with the new PDF Interpreter"

In fact all of these functions do work. The problem (or at least, the
problem I see) is that the original PDF file has a /CropBox entry for
the page, and this is automatically sent (via pdfmark) from the
interpreter to the pdfwrite device.

The result of this is that (when viewed using a consumer such as Adobe
Acrobat) the CropBox is used as a clip, and so the modified content
disappears outside the clip.

So, we need to be able to not send the various Box parameters to pdfwrite.
In fact, we already have a means to do that, and for exactly the same
reason, the Nup device needs to preven the CropBox being sent for the
pages too.

This is done by a device parameter 'ModifiesPageSize, which, along with
ModifiesPageOrder is always set for the Nup device to prevent Box
values, Outlines and Dests being sent.

This commit adds those parameters to the pdfwrite device; by default
they are false so that all the various pdfmarks are sent. A user can
either set them with -d on the command line or via setpagedevice in
PostScript.

When set ModifiesPageSize stops the interpreter passing along the
Box values, and ModifiesPageorder stops it passing along Dests and
Outlines.

Also fixed a typo in a comment, noticed in passing.