Avoid a divide by zero error on printer devices
OSS-fuzz #54638

The fuzzed PDF file ends up setting a page size which causes the width
of the output to be 0. gdev_prn_copy_lines() uses that to determine
how many lines to write by dividing the amount of data by the widht of
the raster line. If the width is 0 we get a divide by zero error.

So only divide by the raster width if the width is non-zero.