GhostPDF - bound check the W array values in an XRefStrm
OSS-fuzz #54436

The PDF file had been fuzzed so that one of the W entries was negative,
which is not valid. This later caused problems when we tried to read
that number of bytes (cast to unsigned) from a file into a buffer which
was sized based on the signed value. That caused a buffer overrun and
subsequent crash.