GhostPDF - check for failure to open files when building font map
Bug #706315 "`ps2pdf -sFONTPATH=. ...` crashes with SIGSEGV"

When enumerating files in a directory in order to build a font map, we
were not checking the result of sfopen but assuming it would succeed and
then trying to read from it.

On Windows at least, the file may already be open exclusively which
results in sfopen returning a NULL, which causes a seg fault when we
try to use it in sgets().

I'm unable to reproduce this problem on Linux so I can't be certain it
is the same problem as in the report, but it seems likely.

Fixed by simply checking the return value for NULL and ignoring the
file if it is. We obviously can't use it as a font anyway so there's
no harm in ignoring it.