GhostPDF - set some default values for type 1 fonts
Bug #705872 see comment #4

Although Firefox is wrong to state that an ExpansionFactor of 0 is 'bad'
it still isn't what it should be in fact. If the font doesn't have an
ExpansionFactor then we should write the default.

The same applies to BlueShift and BlueFuzz. Ideally we should also write
default values for UnderlinePosition and UnderlineThickness, but I can't
see an obvious way to do that, they are not in the same structure.