Revise how the pdfi fontmap code checks a font file exists
Originally, the code assumed that any extra font files were on disk files, and
did not account for the possibility that they might be in the romfs (or
potentially access via the Postscript %os% or similer i/o device).

This revises it to cope with such cases.

Stems from Bug 706267.