Bug 706368: alphabits device declares invalid width/height.
We now assume that every device will give a true width and
height (or set them to zero), so that callers can clip graphic
operations to the extent of the device that is being called.

The alphabits device was breaking this, and giving a much smaller
width/height than was really the case. This caused us to incorrectly
clip.