Bug 706378: Fix infinite loop in shadings.
When we go to fill a patch, we first try to restrict its size to
save us time. We were hitting an infinite loop in here due to
fixed point accuracy.

We know the rectangle that needs to be filled, and we know the
corners of the patch. For some files, the patch can be wildly
outside the rectangle, so spending some time reducing it is
worthwhile.

The routine subdivide_patch_fill does this for us. The logic in
here was slightly flawed, but it was supposed to be:

  Check if a given edge of the patch is completely outside the
  left hand edge of the rectangle. If it is, see if the 'left
  hand half' of the patch is also completely outside of the
  rectangle. If so, keep just the right hand half of the patch
  and retry.

  Do the same for the right hand edge, the top edge and the
  bottom edge.

The infinite loop was caused by inaccuracies in finding the
midpoint of the patch to split into. When the patch x coords
differed by just 1, the midpoint may not converge to a single
value, hence meaning we'd loop forever.

The fix for this is to include a constant (1) in the comparison;
if the patch edge is within 1 of the bounding rectangle, then
subdividing further is not likely to help us!

In investigating this however, it became clear that the existing
algorithm wasn't ideal. Firstly, consider a patch that covers a
huge amount of area, but the midpoint is in the middle of the
rectangle. We can't throw away any 'half' of the patch, so we
fail to reduce the area at all.

We fix this by considering not just 'halving' the patch, but
(in the event that a 'halving' isn't possible), considering
chopping a quarter of the patch off instead.

The second problem is that, for example, we may only be able to
take half the patch off the left after we've taken half the patch
off the right. Accordingly, we put the whole logic in a loop
which we iterate around until we fail to make a change.

This change is enough to avoid the infinite loop.

Unfortunately, due to the presence of a shading in the file that
has such huge coordinates that they are problematic fitting into
fixed point, we spend a lot of time in A_fill_region_floats and
the file still takes far longer to complete than we'd like.

This may be fixable by duplicating the logic in A_fill_region_floats
into a floating point version. This might happen in a later commit.